movies/serializers/countries.py
from rest_framework import serializers
from movies.models.countries import Country


class CountrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Country
        fields = '__all__'
