requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==4.2.1
django-import-export==4.2.0
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
tablib==3.7.0
